requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic==2.10.6; python_version < "3.14"
pydantic>=2.11.7; python_version >= "3.14"
PyYAML==6.0.2
beautifulsoup4==4.13.3
apscheduler==3.11.0
python-dateutil==2.9.0.post0
tenacity==9.0.0
pytest==8.3.4
volcengine>=1.0.216
